third_party/blink/web_tests/virtual/mouse-selection-shadow-dom-disabled/external/wpt/shadow-dom/focus/text-selection-with-delegatesFocus-on-slotted-content-expected.txt

This is a testharness.js-based test.
[FAIL] select slotted text in shadow root with delegatesFocus.
  assert_greater_than: expected a number greater than 0 but got 0
Harness: the test ran to completion.
